chore: Fix and update base image update automation
A previous change to the output from the support script broke the
action; this fixes that and ties to ensure that the commit message will
be properly formatted from this point forward.

This also explicitly upgrades to create-pull-request v7 and enables
commit signing with a GitHub app, which should permit the actions to be
run automatically without requiring modification by a maintainer,
allowing this to be merged faster. As such, the `draft` configuration
has been removed and a step to enable auto-merge has been added.`